Dashboards Mutation API: extract shared dashboardSettingsUtils
Move readDashboardSettings/findCursorSyncBehavior into a shared module instead
of importing across command files, matching the dir's util convention
(layoutPathResolver, annotationUtils, variableUtils, ...).
